requirements.txt
mysql-connector
# git+https://github.com/psycopg/psycopg3.git#subdirectory=psycopg3
psycopg2
psycopg2-binary
tqdm
tracemalloc